+++ src/main/java/ru/skypro/homework/dto/ExtendedAdDto.java
@@ -6,13 +6,13 @@ import java.math.BigDecimal;
 
 @Data
 public class ExtendedAdDto {
-    private Long pk;
+    private Integer pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
     private String email;
     private String image;
     private String phone;
-    private BigDecimal price;
+    private Integer price;
     private String title;
 }
